kafka-cassandra-sink/requirements.txt
cassandra-driver
kafka-python
